implemented function declarations, function pool, and did some refactoring
- added function pool arena for ObjFunction storage
- added function index strong type for type-safe pool references
- implemented OP_CALL with arity checking
- refactored VM to use call frames
- split parser into separate header

+++ include/object.hpp
@@ -1,9 +1,11 @@
 #pragma once
+#include <cstdint>
+#include <string>
 #include "chunk.hpp"
 
 namespace pegasus {
     struct ObjFunction {
-        std::size_t arity_{0};
+        std::uint8_t arity_{0};
         std::string name_{""};
         Chunk chunk_{};
     };
